overlay: test creating lower hardlinks for copied up files
kernel v4.13 introduced the index=on feature for not breaking hardlinks
on copy-up.  With the index feature enabled a regression was introduced -
lower files that are hardlined while overlay is offline can result in
lookup error after overlay in mounted.

The regression was fixed by upstream commit
  6eaf011144af ovl: fix EIO from lookup of non-indexed upper
that was merged to v4.14-rc7.

This test verifies that behavior is sane after creating lower hardlinks
for copied up files while overlayfs is offline.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>