overlay/036: adapt test to run with verify_dir
Both index=on and 'verify_dir' mount options check on mount that upper
dir is not reused with a different lower dir.
Test overlay/037 verifies this behavior for index=on.

In this test however, it is not desired to fail mount on
mismatch of upper/lower, so use the matching upper/lower
dirs in this test.

The mismatch went unnoticed because without 'verify_dir' mount
option, index=off mounts disables this check and index=on mounts
fails on EBUSY (dir in use) before failing on ESTALE (dir mismatch).

Signed-off-by: Amir Goldstein <amir73il@gmail.com>