overlay/036: adapt test to run with incompat index
On kernel compiled with OVERLAY_FS_INDEX_INCOMPAT=y, it is not
allowed to mount an overlay with index=off when there is a non-empty
index dir.

In test overlay/036 there is a sub-test for exclusive upperdir/workdir
that is repeated for index=off and index=on. The sub-test uses a "first"
mount using device name "overlay" and tries to execute two other mounts
that reuse upperdir/workdir respectively.

When running on kernel compiled with OVERLAY_FS_INDEX=y and
OVERLAY_FS_INDEX_INCOMPAT=y, the sub-test fails with index=off, because
the "first" mount was mounted with default index=on.

To fix the test mount all three mounts participating in the sub-test
with same index options. Complement that with a nice helper for the
sub-test and _overlay_check_dirs after every sub-test to verify that
first mount was not corrupted by the other mounts trying to use its
private upper/work dirs.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>