overlay/041: adapt test to 'xino' mount option
The 'xino' overlay mount option is intended to fix inconsistency
of inode numbers in non-samefs overlayfs setup, which is what this
test checks for.

The test is expected to pass on file systems that only support 32bit
inode numbers or if OVERLAY_MOUNT_OPTIONS includes '-oxino'.
Mount non-samefs dirs with OVERLAY_MOUNT_OPTIONS, so user can specify
the '-oxino' mount option when testing with base file system that has
64bit inodes, but is unlikely to use the high inode number bits.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>